Move formatting logic out of QuerySpec
Bug: n/a
Test: n/a

AOSP: c888eca8712e2b62da2b19382f1197462c04d1b9
