+++ querysync/java/com/google/idea/blaze/qsync/query/TruncatingFormattable.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2023 The Bazel Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.idea.blaze.qsync.query;
+
+import java.util.Formattable;
+import java.util.Formatter;
+
+public interface TruncatingFormattable extends Formattable {
+  String toString();
+
+  @Override
+  default void formatTo(Formatter formatter, int flags, int width, int precision) {
+    // We implement Formattable for custom "precision" (max length) handling to allow a truncated
+    // query expression to be shown as such in the log, using normal string formatting.
+    String s = toString();
+    if (precision == -1 || s.length() < precision) {
+      formatter.format(s);
+      return;
+    }
+    final String truncated = "<truncated>";
+    if (precision < truncated.length()) {
+      formatter.format(s.substring(0, precision));
+      return;
+    }
+    formatter.format(s.substring(0, precision - truncated.length()));
+    formatter.format(truncated);
+  }
+}
